refactor: Appの責務分離とカーソル視認性の改善
