+++ src/constants/toolConfig.js
@@ -1,0 +1,4 @@
+export const TOOL_CONFIG = {
+  pencil: { size: 16, composite: "source-over", strokeStyle: "#111827" },
+  eraser: { size: 60, composite: "destination-out" },
+};
